// src/decentium.d.ts
export type Name = string
export type Checksum256 = string
export type TimePointSec = string
export type Uint8 = number
export type Uint16 = number
export type Uint32 = number
export type Uint64 = number | string
export type LinkFlags = Uint8
export type PostOptions = Uint8
export type VariantPlaceholder = ['color', Color] | ['geometrize_triangles', GeometrizeTriangles]
export type VariantMarkNodes = ['link', Link] | ['bold', Bold] | ['italic', Italic] | ['code', Code] | ['strike', Strike]
export type VariantBlockNodes = ['paragraph', Paragraph] | ['heading', Heading] | ['image', Image] | ['linkref', Linkref] | ['quote', Quote] | ['code_block', CodeBlock] | ['divider', Divider] | ['list', List]
export type VariantInlineNodes = ['text', Text] | ['hard_break', HardBreak]
// ...
export interface Bold {
}
export interface CategoryStat {
    endorsements: RewardStat
    linkbacks: RewardStat
}
export interface Code {
}
export interface CodeBlock {
    code: string
    lang: string
}
export interface Color {
    r: Uint8
    g: Uint8
    b: Uint8
}
export interface Divider {
}
export interface Document {
    content: VariantBlockNodes[]
}
export interface FutureExtensions {
}
export interface GeometrizeTriangles {
    base: Color
    triangles: Triangle[]
}
export interface HardBreak {
}
export interface Heading {
    value: string
    level: Uint8
}
export interface Image {
    src: string
    caption: VariantInlineNodes[]
    layout: Uint8
}
export interface ImageInfo {
    width: Uint16
    height: Uint16
    placeholder: VariantPlaceholder
}
export interface Italic {
}
export interface Link {
    href: string
    title: string
}
export interface LinkRow {
    id: Uint64
    from: Permlink
    to: Permlink
    flags: LinkFlags
    payment?: Asset
}
export interface Linkref {
    to: Permlink
}
export interface List {
    type: Uint8
    items: ListItem[]
}
export interface ListItem {
    content: VariantInlineNodes[]
}
export interface Metadata {
    image?: string
    summary?: string
    image_info: PairStringImageInfo[]
}
export interface PairNameCategoryStat {
    key: Name
    value: CategoryStat
}
export interface PairStringImageInfo {
    key: string
    value: ImageInfo
}
export interface Paragraph {
    content: VariantInlineNodes[]
}
export interface Permlink {
    author: Name
    slug: Name
}
export interface PostRef {
    permlink: Permlink
    timestamp: TimePointSec
    category: Name
    options: PostOptions
    tx: TxRef
    edit_tx?: TxRef
    endorsements?: RewardStat
    extensions: FutureExtensions[]
}
export interface PostRow {
    ref: PostRef
    extensions: FutureExtensions[]
}
export interface Quote {
    content: VariantInlineNodes[]
}
export interface RewardStat {
    count: Uint32
    amount: Uint64
}
export interface State {
    moderators: Name[]
    categories: PairNameCategoryStat[]
}
export interface Strike {
}
export interface Text {
    value: string
    marks: VariantMarkNodes[]
}